+++ package.json
@@ -62,7 +62,7 @@
   },
   "scripts": {
     "postinstall": "grunt",
-    "test": "istanbul cover -- node_modules/mocha/bin/_mocha tests/ --reporter spec",
+    "test": "istanbul cover -- node_modules/mocha/bin/_mocha tests/unit/ --reporter spec",
     "start": "node server.js"
   },
   "repository": {
