--- tests/index.js
@@ -1,6 +1,0 @@
-/*global describe, it*/
-describe('website', function () {
-	it('should work', function () {
-
-	});
-});
